Implemented removal of noise in the contract using axioms
Axioms can be provided in a seperate file as input.
Perform conjunction  with each disjunct of a precondition
from final contract and check if it can reduce number of
variables in the disjunct.

Known issue:
There are still clauses with redundant terms.
Perhaps we can look into this in future.